+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 * Upgrade for Dart SDK 1.0
 * Apply new style.
+* Fixed dropdown doesn't trigger close event
+* Fine tune calendar event data
 
 **0.5.2**